+++ simple_iteration/iterations.h
@@ -1,6 +1,8 @@
 #ifndef MNSC_SIMPLE_ITERATION_H
 #define MNSC_SIMPLE_ITERATION_H
 
-double get_next_approx(double r, int iter);
+#include <stddef.h>
+
+double get_next_approx(double r, size_t iter);
 
 #endif //MNSC_SIMPLE_ITERATION_H
